+++ docs/release-notes.rst
@@ -15,6 +15,7 @@ Improvements
 
   * Compatibility with XulRunner/Firefox 30 and 31
   * packages: default Xulrunner version is now 30
+  * setting only with or height of viewportSize is now taking account (#133)
 
 Fixed bugs
 ----------